feat:ai输入框md编辑器内置功能:转为纯文本后按 Markdown 解析
